refactor: move extension method types into Extensions namespace
