+++ src/Xbehave.2.Execution.desktop/ScenarioOutlineRunner.cs
@@ -9,6 +9,7 @@ namespace Xbehave.Execution
     using System.Linq;
     using System.Threading;
     using System.Threading.Tasks;
+    using Xbehave.Execution.Extensions;
     using Xunit.Abstractions;
     using Xunit.Sdk;
