+++ src/Xbehave.2.Execution.desktop/ScenarioRunner.cs
@@ -10,6 +10,7 @@ namespace Xbehave.Execution
     using System.Reflection;
     using System.Threading;
     using System.Threading.Tasks;
+    using Xbehave.Execution.Extensions;
     using Xunit.Abstractions;
     using Xunit.Sdk;
