+++ src/Xbehave.2.Execution.desktop/Xbehave.2.Execution.desktop.csproj
@@ -54,10 +54,11 @@
     <Compile Include="GlobalSuppressions.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Argument.cs" />
-    <Compile Include="ExceptionExtensions.cs" />
+    <Compile Include="Extensions\ExceptionExtensions.cs" />
+    <Compile Include="Extensions\MessageBusExtensions.cs" />
+    <Compile Include="Extensions\MethodInfoExtensions.cs" />
+    <Compile Include="Extensions\TypeInfoExtensions.cs" />
     <Compile Include="Guard.cs" />
-    <Compile Include="MessageBusExtensions.cs" />
-    <Compile Include="MethodInfoExtensions.cs" />
     <Compile Include="ScenarioDiscoverer.cs" />
     <Compile Include="ScenarioOutline.cs" />
     <Compile Include="ScenarioRunnerFactory.cs" />
@@ -68,7 +69,6 @@
     <Compile Include="StepContext.cs" />
     <Compile Include="StepInvoker.cs" />
     <Compile Include="StepRunner.cs" />
-    <Compile Include="TypeInfoExtensions.cs" />
     <Compile Include="Scenario.cs" />
   </ItemGroup>
   <ItemGroup>
